+++ CHANGELOG.md
@@ -78,6 +78,7 @@
 - not truncate ``Error#show`` message having null characters
 - not allow file descriptor number before ``&>``, ``&>|``, ``&>>`` redirections
 - fix typo of builtin ``command`` message
+- fix ``errraise`` option handling in builtin ``command``, ``exec``
 
 ## [0.31.0] - 2023-09-30